+++ R/np.lsqregression.R
@@ -1174,7 +1174,7 @@ nplsqregbw.formula <-
     xdat <- mf[, .np_formula_term_names(attr(attr(mf, "terms"), "term.labels")), drop = FALSE]
     out <- do.call(nplsqregbw, c(list(xdat = xdat, ydat = ydat, tau = tau), dots))
     out <- .nplsqreg_record_omit(out, attr(mf, "na.action"))
-    out$formula <- bws
+    out <- .nplsqreg_retain_formula_terms(out, bws, attr(mf, "terms"))
     out <- .nplsqreg_set_response_name(
       out, .nplsqreg_formula_response_name(bws))
     out$call <- match.call(expand.dots = FALSE)
@@ -1659,20 +1659,17 @@ nplsqreg.formula <-
     ydat <- model.response(mf)
     xdat <- mf[, .np_formula_term_names(attr(attr(mf, "terms"), "term.labels")), drop = FALSE]
 
+    tt <- attr(mf, "terms")
     has.eval <- !is.null(native.exdat) || !is.null(newdata)
     if (!is.null(native.exdat)) {
-      npValidateNewdataFormula(native.exdat, delete.response(tt),
-                               include.response = FALSE)
-      emf <- do.call(.np_formula_model_frame,
-                     list(formula = .np_formula_aligned_terms(delete.response(tt)), data = native.exdat),
-                     envir = parent.frame())
-      eval.omit <- attr(emf, "na.action")
-      exdat <- emf[, .np_formula_term_names(attr(attr(emf, "terms"), "term.labels")), drop = FALSE]
+      prepared <- .nplsqreg_prepare_eval_data(native.exdat)
+      eval.omit <- prepared$omit
+      exdat <- prepared$exdat
     } else if (has.eval) {
       npValidateNewdataFormula(newdata, delete.response(tt),
                                include.response = FALSE)
       emf <- do.call(.np_formula_model_frame,
-                     list(formula = .np_formula_aligned_terms(delete.response(tt)), data = newdata),
+                     list(formula = delete.response(tt), data = newdata),
                      envir = parent.frame())
       eval.omit <- attr(emf, "na.action")
       exdat <- emf[, .np_formula_term_names(attr(attr(emf, "terms"), "term.labels")), drop = FALSE]
@@ -1682,7 +1679,7 @@ nplsqreg.formula <-
 
     bw <- do.call(nplsqregbw, c(list(xdat = xdat, ydat = ydat, tau = tau),
                                 dots))
-    bw$formula <- bws
+    bw <- .nplsqreg_retain_formula_terms(bw, bws, tt)
     fit.args <- list(bws = bw, txdat = xdat, tydat = ydat,
                      gradients = gradients, residuals = residuals,
                      gradient.order = gradient.order, se = se)
@@ -1691,7 +1688,6 @@ nplsqreg.formula <-
     out <- do.call(nplsqreg, fit.args)
     out$call <- match.call(expand.dots = FALSE)
     out$call <- .nplsqreg_describe_call(out$call, parent.frame())
-    out$bws$formula <- bws
     out <- .nplsqreg_set_response_name(out, response.name)
     out <- .nplsqreg_record_omit(out, train.omit)
     out$bws <- .nplsqreg_record_omit(out$bws, train.omit)
@@ -1704,12 +1700,6 @@ nplsqreg.formula <-
       out$eval.nobs.omit <- length(eval.omit)
       out <- .nplsqreg_pad_fit_outputs(out, eval.omit)
     }
-    if (!is.null(out$bws$tau.bws))
-      for (j in seq_along(out$bws$tau.bws))
-        out$bws$tau.bws[[j]]$formula <- bws
-    if (!is.null(out$tau.fits))
-      for (j in seq_along(out$tau.fits))
-        out$tau.fits[[j]]$bws$formula <- bws
     out
   }